restructure and draft complete feature set
